Ready for CI integration!
- Move QUnit tests into one file.
- Add back `junit.js` to serialize results to XML for Jenkins
- Fix a bug in isStale on bjorked cached resources.
