--- tests/enhance.html
@@ -1,47 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <link rel="stylesheet" href="qunit/qunit.css">
-    <script src="qunit/qunit.js"></script>  
-    <!-- <script src="qunit/junit.js"></script> -->
-</head>
-<body>
-    <!-- QUNIT -->
-
-    <h1 id="qunit-header">QUnit Test Suite</h1>  
-    <h2 id="qunit-banner"></h2>  
-    <div id="qunit-testrunner-toolbar"></div>  
-    <h2 id="qunit-userAgent"></h2>  
-    <ol id="qunit-tests"></ol>  
-
-    <!-- FIXTURES -->
-
-
-    <!-- SETUP -->
-
-    <script src="../vendor/zepto.js"></script>
-    <script>Mobify = {$: $, config: {}}</script>
-    <script src="../api/enhance.js"></script>
-
-    <!-- TESTS -->
-    
-    <script>
-
-
-module('enhance')
-
-test('Mobify.enhance', function() {
-    Mobify.enhance()
-
-    ok(1 == 1)
-});
-
-
-// $(window).bind('orientationchange', function() {
-//    alert(Mobify.getOrientation());
-// });
-
-
-    </script>
-<body>
-</html>
