--- tests/externals.html
@@ -1,233 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <link rel="stylesheet" href="qunit/qunit.css">
-    <script src="qunit/qunit.js"></script>  
-    <script src="qunit/junit.js"></script>
-</head>
-<body>
-    <!-- QUNIT -->
-    <h1 id="qunit-header">QUnit Test Suite</h1>  
-    <h2 id="qunit-banner"></h2>  
-    <div id="qunit-testrunner-toolbar"></div>  
-    <h2 id="qunit-userAgent"></h2>  
-    <ol id="qunit-tests"></ol>  
-
-    <!-- FIXTURES -->
-    <script id="test-disable-1" type="text/test">
-        <link href="style.css" />
-        <img src="img.jpg" width="1" height="1" />
-        <iframe src="iframe.html"></iframe>
-        <style></style>
-        <script src="script.js" type="text/javascript"><\/script>
-    </script>
-
-    <script id="test-disable-2" type="text/test">
-        <img src='src.jpg' />
-        <script>
-            "<img src='src.jpg' />";
-        <\/script>
-        <img src='src.jpg' />
-    </script>
-
-    <script id="test-disable-multiline" type="text/test">
-<script 
-    type="text/javascript" 
-    src="https://t.p.mybuys.com/js/mybuys3.js">
-<\/script>
-    </script>
-
-    <script id="test-disable-script-in-comment" type="text/test">
-<!-- <script> -->
-<script 
-    type="text/javascript" 
-    src="https://t.p.mybuys.com/js/mybuys3.js">
-<\/script>
-    </script>
-
-    <script id="test-enable-1" type="text/test">
-        <link x-href="style.css" />
-        <img x-src="img.jpg" x-width="1" x-height="1" />
-        <iframe x-src="iframe.html"></iframe>
-        <style media="mobify-media"></style>
-        <script x-src="script.js" type="text/mobify-script" x-type="text/javascript"><\/script>
-    </script>    
-
-
-    <script id="test-extractHTMLFromElement" type="text/test">
-        <div>
-            <!-- comment -->
-            <script>window.Mobify<\/script>
-            <script src="mobify.js"><\/script>
-            <title>title</title>
-            <plaintext>plaintext</plaintext> 
-        </div>
-    </script>
-
-    <script id="test-extractHTML" type="text/test">
-        <head></head>
-        <body></body>
-    </script>
-
-
-    <iframe id="test-unmobify" src="fixtures/unmobify-basic.html"></iframe>
-    <a href="fixtures-unmobify/basic.html">BASIC</a>
-
-    <!-- SETUP -->
-    <script src="../vendor/zepto.js"></script>
-    <script>
-        Mobify = window.Mobify || {};
-        Mobify.$ = $
-    </script>
-    <script src="../api/externals.js"></script>
-    <script src="../api/extractHTML.js"></script>
-
-
-    <script>
-// iframe setup.
-READY = false
-WAITING = Mobify.$('iframe').length
-
-Mobify.$('iframe').on('load', function() {
-    WAITING--;
-    if (READY && !WAITING) runTests();
-});
-    </script>
-
-
-    <!-- TESTS -->
-    <script>
-
-function runTests() {
-
-
-function getIframeDocument(selector) {
-    return $(selector)[0].contentDocument
-}
-
-
-function getText(selector) {
-    // Fix <\/script>.
-    // Remove extra whitespace.
-    return $(selector).html()
-                      .replace(/\\/g, '')
-                      .replace(/\s+/g, ' ')
-                      .trim()
-}
-
-// http://stackoverflow.com/questions/1304374/javascript-can-a-constructor-function-create-a-documentelement-object
-function createDocument() {
-    var ns = 'http://www.w3.org/1999/xhtml'
-    return document.implementation.createDocument(ns, 'html', null)
-}
-
-(function() {
-    module('externals')
-
-    test('Mobify.html.disable', function() {
-        var got = Mobify.html.disable(getText('#test-disable-1'))
-          , vow = '<link x-href="style.css" /> '
-                 + '<img x-src="img.jpg" x-width="1" x-height="1" /> '
-                 + '<iframe x-src="iframe.html"></iframe> '
-                 + '<style media="mobify-media"></style> '
-                 + '<script type="text/mobify-script" x-src="script.js" x-type="text/javascript"><\/script>'
-
-        equal(got, vow);
-    });
-
-    test('Mobify.html.disable does not escape resources in scripts', function() {
-        var got = Mobify.html.disable(getText('#test-disable-2'))
-          , vow = "<img x-src='src.jpg' /> "
-                + "<script type=\"text/mobify-script\"> \"<img src='src.jpg' />\"; <\/script> "
-                + "<img x-src='src.jpg' />"
-
-        equal(got, vow);
-    });
-
-    test('Mobify.html.disable handles multiline resources', function() {
-        // Don't remove newlines from the test string.
-        var got = Mobify.html.disable($('#test-disable-multiline').html().replace(/\\/g, '').trim())
-          , vow = '<script type="text/mobify-script" x-type="text/javascript" x-src="https://t.p.mybuys.com/js/mybuys3.js"> <\/script>'
-
-        equal(got.replace(/\s+/g, ' '), vow);
-    });
-
-    test('Mobify.html.disable handles <script> in comment', function() {
-        var got = Mobify.html.disable(getText('#test-disable-script-in-comment'))
-          , vow = '<!-- <script> --> '
-                + '<script type="text/mobify-script" x-type="text/javascript" x-src="https://t.p.mybuys.com/js/mybuys3.js"> <\/script>'
-        
-        equal(got, vow);
-    });
-
-    test('Mobify.html.enable', function() {
-        var got = Mobify.html.enable(getText('#test-enable-1'))
-          , vow = '<link href="style.css" /> '
-                + '<img src="img.jpg" x-width="1" x-height="1" /> '
-                + '<iframe src="iframe.html"></iframe> '
-                + '<style></style> '
-                + '<script src="script.js" type="text/javascript"><\/script>'
-
-        equal(got, vow);
-    });
-})();
-
-(function($, HTML) {
-    module('extractDOM')
-
-    test('Mobify.html.openTag', function() {
-        var got = HTML.openTag($("<div id=\"<>\" class='<>'></div>"))
-          , vow = "<div id=\"<>\" class=\"<>\">"
-
-        equal(got, vow);
-    });
-
-    test('Mobify.html.closeTag', function() {
-        var got = HTML.closeTag($("<div></div>"))
-          , vow = '</div>'
-
-        equal(got, vow);
-    });
-
-    test('Mobify.html.extractHTML', function() {
-        var doc = createDocument()
-        doc.documentElement.innerHTML = getText('#test-extractHTMLFromElement')
-
-        var got = HTML.extractHTML(doc);
-
-        equal(got.doctype, '')
-        equal(got.htmlTag, '<html>')
-        equal(got.headTag, '<head>')
-        equal(got.bodyTag, '<body>')
-        equal(got.headContent, '')
-        equal(got.bodyContent, '')
-        equal(got.all(), '<html><head></head></body></html>')
-
-    });
-
-    test('Mobify.html.extractHTMLFromElement', function() {
-        var el = $(getText('#test-extractHTMLFromElement'))[0]
-          , got = HTML.extractHTMLFromElement(el)
-          , vow = '<!--comment--><title>title</title>plaintext</plaintext></div>'
-
-        equal(got.replace(/\s/g, ''), vow);
-    });
-
-    // TODO: This test currently fails b/c the result of `unmobify` is cached
-    //       in extractedHTML.
-    test('Mobify.unmobify', function() {
-        var doc = getIframeDocument('#test-unmobify')
-        Mobify.unmobify(doc)
-        ok(true)
-    });
-
-})(Mobify.$, Mobify.html);
-
-} // End runTests
-
-READY = true;
-if (!WAITING) runTests()
-
-    </script>
-<body>
-</html>
