--- tests/extractDOM.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <link rel="stylesheet" href="qunit/qunit.css">
-    <script src="qunit/qunit.js"></script>  
-    <!-- console.log junit comptabile XML. -->
-    <script src="qunit/junit.js"></script>  
-</head>
-<body>
-    <!-- QUNIT -->
-    <h1 id="qunit-header">QUnit Test Suite</h1>  
-    <h2 id="qunit-banner"></h2>  
-    <div id="qunit-testrunner-toolbar"></div>  
-    <h2 id="qunit-userAgent"></h2>  
-    <ol id="qunit-tests"></ol>  
-
-    <!-- FIXTURES -->
-    
-    <!-- SETUP -->
-    <script src="../vendor/zepto.js"></script>
-    <script>
-        Mobify = window.Mobify || {};
-        Mobify.$ = $
-    </script>
-    <script src="../api/externals.js"></script>
-
-    <!-- TESTS -->
-    <script>
-
-module('extractDOM')
-
-test('ok', function() {
-    ok(1 == 1)
-})
-    </script>
-<body>
-</html>
