--- tests/resizeImages.html
@@ -1,69 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <link rel="stylesheet" href="qunit/qunit.css">
-    <script src="qunit/qunit.js"></script>  
-    <!-- <script src="qunit/junit.js"></script> -->
-</head>
-<body>
-    <!-- 
-    USAGE
-    1. run the test server: node tests/server.js
-    2. localtunnel 1337 (so combo can see the scripts)
-    3. navigate to the localtunnel url
-    -->
-
-
-    <!-- QUNIT -->
-
-    <h1 id="qunit-header">QUnit Test Suite</h1>  
-    <h2 id="qunit-banner"></h2>  
-    <div id="qunit-testrunner-toolbar"></div>  
-    <h2 id="qunit-userAgent"></h2>  
-    <ol id="qunit-tests"></ol>  
-
-    <!-- FIXTURES -->
-
-    <div id="test-resizeImages">
-        <img x-src="http://www.mobify.com/i/phone-tablet.png" />
-    </div>
-
-
-    <!-- SETUP -->
-
-    <script src="../vendor/zepto.js"></script>
-    <script>Mobify = {$: $};</script>
-    <script src="../api/resizeImages.js"></script>
-
-    <!-- TESTS -->
-
-    <script>
-$ = Mobify.$
-
-module('resizeImages')
-
-test('$.fn.resizeImages', function() {
-    var $got = $('#test-resizeImages').resizeImages()
-
-    equal($got.length, 1);
-});
-
-test('Mobify.getImageURL', function() {
-    var got = Mobify.getImageURL('http://test/image.jpg')
-      , vow = '//ir2.mobify.com/http://test/image.jpg'
-
-    equal(got, vow)
-
-    got = Mobify.getImageURL('http://test/image.jpg', {maxWidth: 320})
-    vow = '//ir2.mobify.com/320/http://test/image.jpg'
-
-    equal(got, vow)
-
-
-});
-
-
-
-    </script>
-<body>
-</html>
